// src/API.ts
export type Dish = {
  __typename: "Dish",
  id: string,
  name: string,
  description?: string | null,
  price: number,
  salePrice: number,
  images?: Array< string | null > | null,
  restaurantID: string,
  createdAt: string,
  updatedAt: string,
  _version: number,
  _deleted?: boolean | null,
  _lastChangedAt: number,
};
// ...
export type BasketDish = {
  __typename: "BasketDish",
  id: string,
  quantity: number,
  Dish?: Dish | null,
  basketID: string,
  createdAt: string,
  updatedAt: string,
  _version: number,
  _deleted?: boolean | null,
  _lastChangedAt: number,
  basketDishDishId?: string | null,
};
// ...
export type Basket = {
  __typename: "Basket",
  id: string,
  BasketDishes?: ModelBasketDishConnection | null,
  userID: string,
  restaurantID: string,
  createdAt: string,
  updatedAt: string,
  _version: number,
  _deleted?: boolean | null,
  _lastChangedAt: number,
};

export type ModelBasketDishConnection = {
  __typename: "ModelBasketDishConnection",
  items:  Array<BasketDish | null >,
  nextToken?: string | null,
  startedAt?: number | null,
};
// ...
export type Restaurant = {
  __typename: "Restaurant",
  id: string,
  name: string,
  deliveryFee: number,
  maxDeliveryTime: number,
  minDeliveryTime: number,
  rating?: number | null,
  address: string,
  latitude: number,
  longitude: number,
  adminSub?: string | null,
  icon?: string | null,
  images?: Array< string | null > | null,
  Dishes?: ModelDishConnection | null,
  Baskets?: ModelBasketConnection | null,
  createdAt: string,
  updatedAt: string,
  _version: number,
  _deleted?: boolean | null,
  _lastChangedAt: number,
};

export type ModelDishConnection = {
  __typename: "ModelDishConnection",
  items:  Array<Dish | null >,
  nextToken?: string | null,
  startedAt?: number | null,
};

export type ModelBasketConnection = {
  __typename: "ModelBasketConnection",
  items:  Array<Basket | null >,
  nextToken?: string | null,
  startedAt?: number | null,
};